requirements.txt
gradio
faster-whisper
google-genai
mediapipe
numpy
opencv-contrib-python
python-dotenv
torch
tqdm
yt-dlp
